Reorganize the end-to-end tests around user flows
The one-file-per-assertion layout (correct-flag, wrong-flag,
state-persists, ...) scattered pieces of the same user flow across
files. This regroups them into flow-oriented files with describe blocks:
answering.spec.js (answering flow + session continuity), audio.spec.js
(playback + manual trainer), levels.spec.js, profiles.spec.js, and a
describe wrapper for the existing single note tests.

Pure reorganization: the tests themselves are unchanged and new coverage
comes separately.
